Move sort and filter logic to new module glyph-organizer.js
