+++ src/fontra/client/web-components/glyph-search-field.js
@@ -28,7 +28,7 @@ export class GlyphSearchField extends SimpleElement {
     }
   `;
 
-  constructor() {
+  constructor(options) {
     super();
     this.searchField = html.input({
       type: "text",
@@ -37,9 +37,19 @@ export class GlyphSearchField extends SimpleElement {
       oninput: (event) => this._searchFieldChanged(event),
     });
 
-    this._glyphNamesListFilterFunc = (item) => true; // pass all through
+    if (options?.observer) {
+      this._observer = options.observer;
+      this._observerKey = options.observerKey || "searchString";
+      this._observer.addKeyListener(this._observerKey, (event) => {
+        this.searchField.value = event.newValue;
+      });
+    }
 
     this.shadowRoot.appendChild(this.searchField);
+  }
+
+  get searchString() {
+    return this.searchField.value;
   }
 
   focusSearchField() {
@@ -47,66 +57,16 @@ export class GlyphSearchField extends SimpleElement {
   }
 
   _searchFieldChanged(event) {
+    event.preventDefault();
+    event.stopImmediatePropagation();
     const value = event.target.value;
-    const searchItems = value.split(/\s+/).filter((item) => item.length);
-    const hexSearchItems = searchItems
-      .filter((item) => [...item].length === 1) // num chars, not utf16 units!
-      .map((item) => item.codePointAt(0).toString(16).toUpperCase().padStart(4, "0"));
-    searchItems.push(...hexSearchItems);
-    this._glyphNamesListFilterFunc = (item) => glyphFilterFunc(item, searchItems);
+
+    if (this._observer) {
+      this._observer.model[this._observerKey] = value;
+    }
 
     this.onSearchFieldChanged?.(event);
-  }
-
-  sortGlyphs(glyphs) {
-    // This arguably doesn't belong here
-    glyphs = [...glyphs];
-    glyphs.sort(glyphItemSortFunc);
-    return glyphs;
-  }
-
-  filterGlyphs(glyphs) {
-    return glyphs.filter(this._glyphNamesListFilterFunc);
   }
 }
 
 customElements.define("glyph-search-field", GlyphSearchField);
-
-function glyphFilterFunc(item, searchItems) {
-  if (!searchItems.length) {
-    return true;
-  }
-  for (const searchString of searchItems) {
-    if (item.glyphName.indexOf(searchString) >= 0) {
-      return true;
-    }
-    if (item.codePoints[0] !== undefined) {
-      const char = String.fromCodePoint(item.codePoints[0]);
-      if (searchString === char) {
-        return true;
-      }
-    }
-  }
-  return false;
-}
-
-function glyphItemSortFunc(item1, item2) {
-  const uniCmp = compare(item1.codePoints[0], item2.codePoints[0]);
-  const glyphNameCmp = compare(item1.glyphName, item2.glyphName);
-  return uniCmp ? uniCmp : glyphNameCmp;
-}
-
-function compare(a, b) {
-  // sort undefined at the end
-  if (a === b) {
-    return 0;
-  } else if (a === undefined) {
-    return 1;
-  } else if (b === undefined) {
-    return -1;
-  } else if (a < b) {
-    return -1;
-  } else {
-    return 1;
-  }
-}
